将原有的视频点赞模型从 VideoUpvoteSchema.ts 迁移到 VideoVoteSchema.ts 新增视频点踩模型 VideoDownvoteSchemaFactory
